requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
wtforms_components
flask_restful==0.3.5
RPi.GPIO
apscheduler
redis
flask-socketio
aiohttp
aiohttp_jinja2
